Renamed GroupInput component to MultiCheckbox
