+++ src/components/FormInputs/index.ts
@@ -1,6 +1,6 @@
 export { default as ColorInput } from './ColorInput.svelte';
 export { default as Checkbox } from './Checkbox.svelte';
 export { default as Dropdown } from './Dropdown.svelte';
-export { default as GroupInput } from './GroupInput.svelte';
+export { default as MultiCheckbox } from './MultiCheckbox.svelte';
 export { default as NumberInput } from './NumberInput.svelte';
 export { default as TextInput } from './TextInput.svelte';
